requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cloudinary==1.26.0
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2-binary==2.9.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
